Fixed behavior with sub-images accessed through same ImageServer
Getting a suitable ImageData path within a project previously depended
on the ProjectBrowser class.  Now it depends on QuPathGUI instead,
which seems at least a bit more sensible.

This change was part of a wider revision to prevent
PathImageDetailsPanel from going rogue with its auto-saving of
ImageData (which were introduced as a temporary measure before Projects
were created, but managed to hang on far too long through by keeping a
sufficiently low profile).  This improves sensibleness in the case when
multiple images occur within the same file.