Image tab & stain vector setting fixes
Improvements when manually setting stains.  More meaningful message,
including if stains are invalid (i.e. two are equal and a residual is
required).

Text colors in the table within the image tab are now set properly;
previously they could be adjusted by color deconvolution stains, and
not reset to black if a cell was reused.