ImageData & fluorescence data performance improvements
Changes to improve how brightness/contrast settings are managed.

These are now stored (as a JSON property) in serialized ImageData
objects.  Additionally, when repainting a serialization of these
properties is also used to construct a key for RGB tile caching.  This
is applied for image types that are expensive to paint (i.e. > 1
channel, non-RGB) because of the color transforms involved.

A small extra change: .qpdata files in a project are no longer saved
automatically if there have been no changes & no .qpdata file already
exists.  This makes the presence or absence of a .qpdata file more
informative.

Adjusting the brightness/contrast now counts as a ‘change’, and a
prompt will be added asking whether to save or not.

Finally, the auto brightness/contrast setting has been made less
extreme.