Storing available PathClasses in projects
The list of available PathClasses is now stored in the .qpproj file, to
help with setting up different classification lists for different
projects.

On right-click of the classification list, it is also possible to
import the classifications from a different project.

Toggling the display of a particular PathClass on/off is also now
indicated in the displayed PathClass list.