Text descriptions can be added to annotations
Text descriptions can be stored for annotation objects.

These are editable along with properties.  To make them easier to
access, the shortcut ENTER can be used to open properties for a
currently-selected annotation (similar to adding notes for a TMA core).

Previously ENTER would cENTER the object instead; this still happens
for detections.