Fixed bugs when no pixel sizes set
Fixed bug that prevented intensity measurements from being generated
whenever a square or circular tile size was selected for an image with
no pixel size information available.

Also fixed bug that resulted in centroids being shown as NaN if no
pixel size was set.

Also reduced logging priority when adding an object to the hierarchy
(since it produced a lot of messages…).