Improved stain setting & relaxed cytokeratin H-DAB rule
Creating cytokeratin annotations no longer performs (excessively-strict) check that the image is identified as H-DAB.  This means it can be more easily repurposed for other ROI-generating purposes.

The actual processing is unchanged so long as the third stain is a residual stain vector; Stain 2 is consideredto be the main marker (i.e. DAB previously), while the tissue is the sum of either Stain 1, Stain 2 and Stain 3 (if it is not a residual stain vector), or Stain 1 and Stain 2 (if Stain 3 is a residual stain vector).

The parameter keys are also retained, i.e. thresholdDAB is still the key even if the stain itself isn't actually DAB.

When setting stain vectors, the option to change the name is now only enabled for 'Brightfield (Other)' to prevent confusion; also, the option to specify the collective names for the stain vectors has been added.