Greatly reduced use of ImageRegionStore
This was previously the main mechanism to make use of caching when making many requestions, relied upon by several plugins (e.g. for texture features).

However, the aim now is to make caching internal when requesting from (some) ImageServers, so that *all* requests benefit, not only those that go through an ImageRegionStore.

This should improve performance (or at least not cause it to be too much worse), while substantially simplifying the code.