Added BufferedImageOverlay & revised PathOverlays in general
The BufferedImageOverlay makes it possible to add simple pixel-based overlays on top of one or more image regions.

Overlays are now split into custom & core overlays (with the latter always painted on top), and accessible via ObservableLists.  This makes it easier to control the contents of the overlays, and potentially to track them.

Some extra changes in overlay behavior also, so that when opening a new image an overlay that returns false for supportsImageDataChange() will now automatically be removed from the viewer.