Cleaned up color deconvolution code
Reorganized much of the code where color deconvolution was implemented, removing methods that were not essential.  This makes it easier to maintain the code in the future, and to ensure that color deconvolution is applied consistently (e.g. using 'white' values, and not defaulting to 255).

While making these changes, unnecessary and legacy nucleus/membrane detection algorithms were also removed.