Added dumpMetadata() method to Openslide/ImageJ servers
This makes it possible to request a string representation of whatever metadata the ImageServer has access to. This helps to check whether there are problems parsing on the QuPath side, or if the actual metadata has an issue.

The way it's returned is not defined (e.g. JSON, XML... or something else).

Also modified the default color for the 6th channel (which was wrongly a duplicate of channel 4).