Fixed PathClass.compareTo method
`PathClass.compareTo` is now equivalent to
```
this.toString().compareTo(o.toString())
```

Note that in previous versions (< 0.1.2), the comparison was made based on the name only.

This could result in unexpected behavior whenever comparing with equality and using derived `PathClass` objects, because only the (final) name part was being compared and this could potentially result in classifications (wrongly) being considered equal (e.g. "Tumor: Positive" and "Stroma: Positive").

This was most significant when working with Groovy, where `==` is replaced by `compareTo`.