Changed method of displaying counts per classification
Annotation/TMA measurements of the form 'Num Class A' now give counts of the number of objects with exactly the classification 'Class A' - and not either 'Class A' or any derived classification, as was previously the case.

This change is made to help support multiple classifications with minimal disruption.

Where intensity classifications are used, it can still be useful to have the counts of all objects with the same base classification - with the intensity part stripped off.  In this case, the new corresponding column heading would be 'Num Class A (base)', along with 'Num Class A: Positive' and 'Num Class A: Negative' (for example).