Reduced visibility of ROI constructors
Introduced ROIs class containing static methods to construct different ROIs, rather than depending on constructors.

This is a step towards recaliming more control (on the QuPath development side) regarding the kinds of ROIs that can be created, which can in term help ensure that these may all be painted/converted as required.

Through the refactoring, it is clear that there is still too much reliance on default planes (c, z, t), and that PointsROIs require a more flexible method of handling different planes.