Revised MiniViewerCommand, thumbnail generation and histograms
Rewrote MiniViewerCommand so that it no longer creates subclasses of QuPathViewer, but rather uses a custom Canvas component.  This should be much more efficient, and avoid the need to handle all the pain that can come from setImageData in QuPathViewer (which impacts the ImageDisplay etc.).

Among the other benefits, launching the channel viewer is much faster.  Painting is also now performed in a background thread to further improve performance.

Other changes:
- Revised histograms to make them immutable (previously could set normalized status)
- Added getDefaultThumbnail methods to ImageServer
- Revised ChannelDisplayInfo to remove setters from core interface
- Rewrote histogram-caching mechanism used with ImageDisplay
- Removed requirement for ImageRegionStore within ImageDisplay
- Removed SecondViewerOverlay (which wasn't useful)
- Added option to access last repaint timestamp in QuPathViewer