Updated use of StatModels
First (semi-working) effort at updating the use of OpenCV StatModel for classification.

Idea is to make StatModels serializable to JSON, with a view to combining this with other required classification information (e.g. feature names) - both for object and pixel-based classification.

Additionally, incorporated option to apply PCA for feature reduction and to log 'variable importance' where calculated for RTrees.

Currently, PCA is always on... option to adjust it needs exposed in the GUI.

TypeAdaptersCV gives the primary mechanism to use OpenCV's serialization to create a JSON string for a StatModel, Mat or SparseMat.

Also fixed bug that prevented classification probabilities being updated consistently along with classifications.