Added always paint selected objects preference
Added option to always paint selected objects, even when the overlay opacity is zero.  This handles the annoying situation where an image is being annotated but with a transparent overlay, and on completion of the annotation it disappears.

Also modified the calculation of ROI handle sizes to improve the situation for Polylines.