Updated to JDK 11.0.2 and fixed test
Updated to JDK 11.0.2, which fixes a bug on macOS when modal dialogs regained focus.

Also turned off GPU requirements for OpenCV in build script (needed to work within eclipse on macOS?).

Adjusted a hierarchy test because of recent change in the event that fired when an object without children is removed (now a remove event, previously a change event).