Changed polygon and polyline tool behavior
Polygon and polylines are (optionally) now 'freehand' if they begin with a drag operation; otherwise they are based on clicked points (perhaps with freehand components).

Basically this means that if a ROI is drawn by clicking and dragging (exlusively), then it is completed when the mouse button is released - rather than requiring a double-click later.