Removed internal buffer when applying ImageDisplay transforms
The buffer became a liability when using multithreading, specifically when using the Channel Viewer.
Now a new float array is created when requesting values.  It remains to be seen if this is sufficiently noteworthy as a performance issue to require an API change (e.g. optionally providing a reusable float array to updateRGBAdditive).