Fixed tile caching and move tool bugs
Fixed bug that meant the wrong tile could be provided by an AbstractTileableImageServer, because the path element of the request was ignored.

Also changed the event fired when an annotation is modified with the MoveTool, allowing it to count as an 'added' event (since the object is being re-added to the hierarchy).  This means that the object itself is available for querying in the event, which makes it easier for other commands to respond appropriately (e.g. interactive classifiers).