Pixel classifier improvements
Pixel classifiers now require an ImageData, not merely an ImageServer.  This enables them to (potentially) use other information, such as color deconvolution stains or the image type.

RGB tile caching is applied less often in the PixelClassificationOverlay; for classified results it is not necessary for performance, and only consumes memory.  Additionally, the painting code has been rearranged so that the RGB cache tile might (in the future) outlive the original tile, which need not remain cached if memory is at a premium.  Relatedly, a memory estimate is logged when the classifier server is created.

Fixed (kind of) a bug whenever ellipse annotations were used to mask the PixelClassificationOverlay; now uses ellipse bounding box.

Delayed initialization of StatModels until they are required; this means that the Pixel Classifier window opens _much_ faster.

Removed unnecessary RGB/BGR switches in OpenCVPixelClassifierDNN, and now set it to set a channel to zero if scale is zero (it would probably be wise to change scale to a multiplication rather than a division).