Lines, points & polylines can be selected by double-clicking
Lines, points & polylines can be selected by double-clicking on the image when the Move tool is selected, rather than needing to select them some other way.

The previous rule was that the ROI area needed to contain the clicked point... which had obvious limitations for ROIs that did not enclose an area.