PathObject changes
Switched again from TreeSet to use LinkedHashSet for child object list.

The 'trouble' with TreeSet was its use of the compareTo method to resolve whether or not to add an object; if this returned zero, only one object would be included. This may be a good thing, unless the comparator is not sufficiently reliable.  Some comparator changes have been made, but because it was only intended to give a rough ordering for reproducibility when listing objects, it is not considered reliable enough for this purpose.

Tests have been updated to assess if child list contents are the same, rather than true equality for a list or collection (which behave in different ways).  The order should be considered undefined.

Improved ConverterJTS.