Improved pixel classification measurement performance
Measurements are performed in a background thread, and pixels are accessed in bulk from a ThreadLocal mask for each tile - as this was previously a bottleneck with large numbers of images.

Coordinates when generating the masks are now made using TileRequests, which allows for improved accuracy compared to using RegionRequests as before.

isLocked and setLocked are now methods for all PathObjects, so as to reduce the need to determine specifically if an object is e.g. a PathAnnotationObject instance.

doSplit defaults to false, since splitting slows down the object creation.