Revised writing classified tiles and ImageServer metadata
Pixel classification now only writes classifiers & tiles when the 'Save and apply' button is pressed.

This also includes writing image metadata, but the reliance on pixel sizes in microns didn't look great.  The metadata has now been revised to introduce a PixelCalibration object, with a view to (eventually) supporting alternative units for the pixel size.