Updated ImageServers to use URIs
ImageServerBuilders now use URIs rather than String paths, which makes the scheme clear and helps test whether support is available more effectively.  Bio-Formats servers have been revised to take the series number as a fragment, reducing the reliance on a name being present. For backwards compatibility, a name can still be provided as a query.

The cache is no longer provided as a parameter to the constructor, but rather is taken automatically from the ImageServerProvider by any server that requires it (currently all AbstractTileableImageServers).

ImageServerProvider can now optionally take an array of preferred builders, to give control over preference.