Reading and writing ImageData now throws exceptions
In a good way, that is (more or less). Previously the exception would be lost within PathIO, and the first one would know about it was when null came back.

DefaultProjects now also write server metadata to json & read it again when required, finally giving the possibility of persistent metadata changes (e.g. to set wrong pixel sizes or channel names).