Saving and reloading new style projects
Projects based on 'uniqueName' rather than 'imageName' can now be saved and reloaded.

Thumbnail storage has been moved inside the projects, and the mechanism to add images to the project has been revised to generate the thumbnail at the time when the image server is first added (rather than at a later date). This can be more efficient and also fail earlier if the image is not readable.