Revising projects
Changes to reduce the reliance on projects being based on the local file system. Improved consistency by trying to avoid having a project and image open simultaneously but where the image is not itself part of the project.

Improved use of image names when images read by Bio-Formats have multiple sub-images.