Improved tile size when writing images, fewer log messages
OMEPyramidWriter now uses the full image as the tile size whenever the server does the same, which is especially important for images with many channels (writing performance was previously extremely slow).

Some log messages have been relegated to debug so as to be less frequent and distracting.

Added new GeneralTools.deleteFile(boolean) method to request sending to trash instead; this requires Desktop to support the operation.