ImageServer fixes and new ConcatChannelsImageServer
Added warnings whenever an ImageServer resizes tiles in unexpected ways, and added ability to extend (not only crop) a tile if the request cannot be fulfilled directly. This fixes a problem seen when combining a SparseImageServer and CroppedImageServer, where tiles would be resized wrongly.

Added ConcatChannelsImageServer to dynamically append the channels from other ImageServers.