Fixed server building bug and polygon typo
The use of a TreeMap meant that if two ImageServerBuilders had the same support level, only one would have a chance to open the file - which thwarted Bio-Formats being able to handle files that it was otherwise capable of (e.g. if ImageJ got to try instead).

Fixed unfortunate polygon typo in ROIs class.