Return Collection not List when requesting objects from hierarchy
A small fix that got out of hand... there was a performance issue checking 'containsAll' for a list of annotations in PathAnnotationPanel. Now it's possible to specify objects should be in a set rather than a list (for example).

Alternative would have been possible just add everything to a set afterwards... but potentially better to generalize to all collections instead.

Also updated jts and made visible within GUI.