Selection model changes
Selection change events now include the full list of changed objects (not just the primary object), so can be handled without needing to return to query the selection model itself.

Major change to hierarchy & annotation panels to synchronize selections only to the primary selection. The aim is to remove a performance bottleneck.
Additionally, synchronization is not performed if requested from another thread. The aim is to avoid potentially thwarting scripts by causing selection changes through the GUI at a later time than required within the script.