Improved BioFormatsImageServer multithreading
Reduced the complexity of the overly-complicated (and probably leaky) approach to threads when using BioFormatsImageServers. Now we rely on IFormatReaders (no longer BufferedImageReaders) created statically per-thread rather than with a further dependency on the specific ImageServer.

The aim was to place limits on the number of ImageReaders active at any one time, and to improve cleanup whenever threads may be stopped.