Do image checks earlier in BioFormatsImageServer
Check image validity while initializing a BioFormatsImageServer to reduce the chances of adding invalid images to the available image list, which will then later fail. Also, default to largest image in a series only if it is substantially larger than the first image - otherwise stick with the first one.

Changes required to deal with some czi TMA files where the macro image (wrongly) was identified as being huge, which then caused out of memory errors if read in one go, or alternatively array index exceptions if trying to switch resolutions. Now this image should be identified as invalid and not displayed to the user.