Project entries indexed by ID
Project entries now indexed by ID rather than path. The primary reason is that paths are less reliable and/or harder to calculate whenever servers may be complex and wrap other servers.

The ID is stored as a properly of the ImageData, and inserted whenever it is being read from a project.

A consequence of this is that it's easier to add duplicate images to a project, since there is no longer a search to check if an image with the same path already exists.