Serialize QuPath ROIs to and from GeoJSON
Added code to convert between QuPath ROIs and GeoJSON, via JTS Geometry objects.

As part of this, fixed a longstanding bug that meant lines and polylines were not properly handled by getShapeROI, and Geometry objects representing LineStrings and Points/Multipoints were also not correctly converted to QuPath ROIs.

Introduced GsonTools, which provdies a static Gson instance with type adapters registered to write key QuPath classes. Currently this only includes ROIs, but the plan is to expand this and avoid the need to construct separate Gson instances in various places.