Removed ID from ImageServerMetadata
The ID was previously used for the path, but since it was derived from the URI then a find & replace applied to the project JSON would start manipulating paths... which is undesirable. It also opened the possibility of setting the path to something non-unique within the project.

Paths are now generated differently, and typically include the classname for the server (making uniqueness easier to achieve). Their primary purpose is for tile caching.

The Image details panel shows URIs rather than the path, because the path is typically no longer very readable or informative.

Also fixed NPE in viewer when navigating with arrow keys and not having a TMA grid.