Updated build script
Removed 'create-app-image' and 'create-installer' options, which are no longer supported in jpackage build 8.

Because executables are now inside the 'bin' directory of images, and are potentially harder to find, updated to create .msi installer for Windows. Installation must be local to support .cfg file changes to memory limit on launch.

However, also added -XX:MaxRAMPercentage=50 Java option to set memory default to 50% available memory, making it less crucial to change the setting on most systems.

Added a secondary console launcher on Windows to help with debugging.

Updated ij and license report dependencies.