Reuse IFormatReader where possible
Bio-Formats servers now reuse the same primary IFormatReader if possible, and rely on Cleaner objects to close these when no longer required by any active servers.

This may not be quite as nice as each server having its own reader and closing this as required, but it is more convenient. There were also severe performance issues for readers that were slow to initialize (e.g. .czi file of a TMA with many scenes).