Improved tile caching
Switched to use a Guava cache, optionally setting the percentage of available memory used for tile caching in the preferences.

The hope is that this (and use of soft values) will reduce out-of-memory errors by better enforcing the cache size, while giving the opportunity to use larger caches where RAM permits. This is especially important for multiplexed images.

This means that care needs to be taken when returning null from readBufferedImage(); ideally null would not be returned since it is not a supported value in the cache.

Introduced 'empty' tiles within AbstractTileableImageServer to mitigate this. This allows sparse/pyramid-generating servers to return the same tile for multiple requests if there is nothing present (i.e. an unscanned region). This is not stored in the main cache, but rather locally within the server. The memory requirements are considered negligible.

This can dramatically reduce the number of pixel arrays being stored for sparse images, and thereby improve the usefulness of the cache.

Fixed a bug caused by delayed loading of Wand tool, where an exception was thrown if the Preferences were opened before the Wand was available.