Update to JavaFX 13, improved measurement export and object selection
- Improve toString() method of ROIs and ImagePlane
- Include image name & annotation count in measurement tables
- New saveImageMeasurements scripting command
- Write measurement tables one row at a time (rather than creating a single mega-String); show an exception if attempting to create a String that is too large
- Add PathObject.isCell() method
- Fix synchronization when requesting PathClasses (necessary because subcellular detection through exceptions due to initializing a pre-existing class)
- Simplify code for synchronizing some UI elements with PathObject selection status

It appears that JavaFX's multiselection models (based on bitset) are horrendously slow when selecting many entries (low thousands). This affects tables and treetables (and possibly lists). The SummaryMeasurementTable will no longer synchronize more than 1000 selections for that reason, but this is insufficient to solve the problem - and directly selecting from the table still results in poor performance.