Improve performance when drawing constrained ROIs
When drawing ROIs constrained not to overlap existing annotations, resolve overlapping annotations when only if required (and not every time a ROI is selected or created). This makes QuPath object selection much more responsive.

Previously, double-clicking to select objects on images with many annotations could fail, because the delay introduced when processing the first click meant the second click was processed to late.