Always write project on exit
This is needed to ensure some changes persist, e.g. if setting channel names for an image but not otherwise changing the project.

Project writing revised to be a bit more cautious: write to a new file, then perform move operations to rename. This attempts to avoid the risk of inadvertently corrupting a project file if writing fails for any reason. It also avoids the need to serialize the JSON to a string in advance.

Also write server.json for each entry. This gives a potential way to rebuild (most of) a project from the data directory alone in an emergency.