Optionally fix slider range to text field input, update density maps
Updated methods in GuiTools to deal with slider/textfield bindings where the user inputs values that are out of range: optionally expand the range and/or support adding an event listener for the user to input a required range.

Density maps updated to handle multiple viewers & auto-update as objects change on the current image.