Propagate script exceptions & use for exit code
Proposed fix for https://github.com/qupath/qupath/issues/654

Note that from the command line the exit code should be 0 either if a script succeeded or if it threw an exception when applied to a single image.

If an exception is thrown when attempting to apply the script to multiple images in a project, QuPath will attempt to recover and continue processing the next image.