Add JavaFX to module path
See https://github.com/openjdk/jfx/blob/jfx16/doc-files/release-notes-16.md#important-changes

This change puts JavaFX where it wants to be, but can potentially cause problems - so more testing is needed. Specifically, there can be trouble where reflection is accessing internal JavaFX classes.

When issues are found, these can generally be worked around with --add-opens and --add-exports - but we need to be on the lookout for any that either QuPath or its dependencies rely upon. And ideally we wouldn't require access to anything that isn't exported.

In the longer term, QuPath itself should become modular - but this remains difficult to achieve because of its non-modular dependencies.