Fixed Gson type adapter regression
ImageOps test would fail intermittently because of the order in which GsonBuilder.registerTypeAdapterFactory(factory) was called.

This could result in a null factory being added, since GsonTools was requesting a factory from ImageServers before it had been initialized.

The currently workaround has all tests passing, but remains brittle with a risk that the issue could resurface with new type adapter factories being registered in the 'wrong' order - particularly if they are added within the static initializer of GsonTools.