Add read only mode to QuPathGUI
This provides a mechanism to suppress 'save' dialogs if QuPath is being run simply as a viewer but without the ability to save changes to any image.

Note that this minimal implementation is for feedback; it may need to be changed in a future version if it proves insufficient in practice.