Update to Gradle 7
This should allow QuPath to build with any JDK from 8-16 (downloading 16 if needed).

Minor fixes needed to ensure the build is with the target version, and avoid including licenses for subdependencies if they are not in the build.

Infer modularity now needs to be turned off to avoid errors in eclipse.

Bump dependencies ImageJ and RichTextFX