Optional args on import, Bio-Formats dimension swapping
Support specifying optional (currently space-delimited) arguments in the import dialog for images.
Enable DimensionSwapper with Bio-Formats.
Putting these together, one can use a String such as
  --swap XYTZC --series 2
to specify that only the third series should be read from a file with Bio-Formats, the dimension order should (possibly) be changed, and autostitching (for Zeiss) should be turned off.