Store JSON rather than an image path in .qpdata files
This makes it possible to reconstruct complex servers more reliably, by replicating the information stored within a project.
This is necessary for things like rotation, color transforms etc. in addition to preserving metadata.
Previously, such servers *had* to be in a project, since the necessary info wasn't encoded within the .qpdata file.