Allow the ImageServerBuilder to be specified when importing images
This means that when importing images from OMERO (for example), the user is not provided the option to specified a different server - and import should work faster, since it doesn't need to check builders that are doomed to fail anyway.

This involved making promptToImportImages non-public, since the methods in ProjectCommands should be used anyway.