Use Mat.close() not Mat.release()
Aims to fix https://github.com/qupath/qupath/issues/753

Use of Mat.release() seems to be a misunderstanding of how deallocation is handled in JavaCPP/OpenCV (or a throwback to the pre-JavaCPP days). It seems that Indexer.release() is valid, but Mat.release() is unhelpful.

Probably most meaningful change is the use of PointerScope in OpenCVPixelClassifier, since this could result in particularly high memory use.