Update image names in both project & server metadata
Fixes https://github.com/qupath/qupath/issues/593
When the image name is set for an image open in the viewer, the metadata is update immediately. Otherwise, the name is set only in the project (as before) and update in the server as entry.readImageData() is called.
Potentially, getting an image some other way (i.e. via a serverBuilder) could fail to give the updated name.