Avoid updating hierarchy/annotation tabs when not visible
This aims to improve performance when running commands on images that already contain many objects. Updating the hierarchy tree/annotation list could have a severe performance impact when triggered for every hierarchy or selection update.

This change makes it possible to block frequent update events whenever the corresponding tab is not selected (and therefore the component is not visible).