Optionally set LD_LIBRARY_PATH for Linux
This avoids the need to create a custom launch script as described at https://github.com/qupath/qupath/issues/628

While still not ideal, it gives an alternative way to avoid using system shared libraries with OpenSlide - which is known to be problematic because of pixman issues on Ubuntu LTS releases (including 20.04).

This works by including a bash script in the jpackage image generated for Linux by default. It also works with
`./gradlew run -Pld-path=true`
and with
`./gradlew installDist -Pld-path=true`