Moved project entry thumbnail generation to background
Attempt at moving project entry thumbnail generation to background (and only process the project entries visible to the user). As of this commit, there is a delay when refreshing the TableView, which makes it slow to navigate. E.g. On mouse click, all the visible rows in the tableView are refreshed, meaning that the thumbnail is re-painted onto the Canvas, which can be slow when a lot of images are visible.