Update gradle wrapper, JavaFX, build scripts
JavaFX update is included based on changelog, since it may fix issues such as https://github.com/qupath/qupath/issues/833

Build scripts now have *extremely* provisional and experimental support for Apple Silicon. To use this, it is necessary to build with a suitable aarch64 JDK and (probably) skip specifying a toolchain, e.g. with
```
./gradlew build jpackage -Ptoolchain=skip
```
Note 1: This will break OpenSlide support. Whole slide images need to be read some other way (e.g. Bio-Formats).

Note2 : This relies upon a JavaCPP-Presets snapshot for OpenCV, so is subject to change and not intended for any serious work. Nevertheless, it can be used to give an impression of performance improvements.