Support for save file choosers that accept multiple extensions
Previously, only a single file filter could be specified. This meant it wasn't possible, e.g. to provide a single chooser to save from a list of image types.

Multi-part file extensions are slightly problematic, at least on macOS, since the chooser will automatically add only the file part of the specified extension. The QuPathChooserFX will attempt to correct for this, but it means that sometimes the file returned is not identical to the one specified in the chooser.