Update dependencies
This includes Bio-Formats 6.9.0 for further testing. It still has an issue with some svs files, but adds dicom support - both should be documented if this remains in the next release.

It also includes JavaFX 18, which fixes a bug that meant selecting all entries in a large table (e.g. a detection measurement table) was unusably slow.