Add a UUID to each PathObject
This is a fairly major change to incorporate IDs for objects.

The purpose is to make it much easier to match up objects whenever some further analysis is done elsewhere (e.g. classification or clustering in Python).

I've attempted to ensure both forwards and backwards compatibility by storing the ID in the metadata map. This is somewhat inefficient, but means that data files can be reopened in earlier QuPath versions. A future QuPath release might change this to a more efficient representation.