BUG FIX: turn off inter-value interpolation for labelmaps
For a labelmap or any image where pixel values mean something specific (e.g. gray values in a CT) using any interpolation that creates values outside the range that already exists is invalid, so labelmap images should use nearest-neighbours interpolation when resampling instead of OpenCV's INTER_AREA, which is the current default for interpolation. This change implements this. 
338890